backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.1
apscheduler==3.10.4
twilio==8.10.0
openai==1.3.5
pydantic[email]==2.5.0
python-multipart==0.0.6
requests==2.31.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
